helios/requirements-test.txt
pytest==2.6.4
pytest-cov==1.8.1
pytest-flask==0.6.2